requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.0.0
notion-client==2.2.1
openai==1.12.0
anthropic==0.18.1
APScheduler==3.10.4
